requirements.txt
nfcpy==1.0.4
pyserial
